last version, problem with logging
